Saving changes serially in SaveHandler
Resolves #96 Technical label not set in one specific case
